objects: move Tree into tree.py
