+++ dvc/objects/__init__.py
@@ -1,10 +1,6 @@
 import itertools
-import json
 import logging
-import posixpath
 from concurrent import futures
-
-from funcy import cached_property
 
 from dvc.exceptions import DvcException
 from dvc.progress import Tqdm
@@ -73,132 +69,9 @@ class HashFile:
         return odb.get(hash_info)
 
 
-class Tree(HashFile):
-    PARAM_RELPATH = "relpath"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._dict = {}
-
-    @cached_property
-    def trie(self):
-        from pygtrie import Trie
-
-        return Trie(self._dict)
-
-    @property
-    def size(self):
-        try:
-            return sum(obj.size for _, obj in self)
-        except TypeError:
-            return None
-
-    def add(self, key, obj):
-        self.__dict__.pop("trie", None)
-        self._dict[key] = obj
-
-    def digest(self):
-        from dvc.fs.memory import MemoryFileSystem
-        from dvc.path_info import PathInfo
-        from dvc.utils import tmp_fname
-
-        memfs = MemoryFileSystem(None, {})
-        path_info = PathInfo(tmp_fname(""))
-        with memfs.open(path_info, "wb") as fobj:
-            fobj.write(self.as_bytes())
-        self.fs = memfs
-        self.path_info = path_info
-        self.hash_info = get_file_hash(path_info, memfs, "md5")
-        self.hash_info.value += ".dir"
-        self.hash_info.size = self.size
-        self.hash_info.nfiles = len(self)
-
-    def __len__(self):
-        return len(self._dict)
-
-    def __iter__(self):
-        yield from self._dict.items()
-
-    def as_dict(self):
-        return self._dict.copy()
-
-    def as_list(self):
-        from operator import itemgetter
-
-        # Sorting the list by path to ensure reproducibility
-        return sorted(
-            (
-                {
-                    # NOTE: not using hash_info.to_dict() because we don't want
-                    # size/nfiles fields at this point.
-                    obj.hash_info.name: obj.hash_info.value,
-                    self.PARAM_RELPATH: posixpath.sep.join(parts),
-                }
-                for parts, obj in self._dict.items()  # noqa: B301
-            ),
-            key=itemgetter(self.PARAM_RELPATH),
-        )
-
-    def as_bytes(self):
-        return json.dumps(self.as_list(), sort_keys=True).encode("utf-8")
-
-    @classmethod
-    def from_list(cls, lst):
-        from dvc.hash_info import HashInfo
-
-        tree = cls(None, None, None)
-        for _entry in lst:
-            entry = _entry.copy()
-            relpath = entry.pop(cls.PARAM_RELPATH)
-            parts = tuple(relpath.split(posixpath.sep))
-            hash_info = HashInfo.from_dict(entry)
-            obj = HashFile(None, None, hash_info)
-            tree.add(parts, obj)
-        return tree
-
-    @classmethod
-    def load(cls, odb, hash_info):
-
-        obj = odb.get(hash_info)
-
-        try:
-            with obj.fs.open(obj.path_info, "r") as fobj:
-                raw = json.load(fobj)
-        except ValueError as exc:
-            raise ObjectFormatError(f"{obj} is corrupted") from exc
-
-        if not isinstance(raw, list):
-            logger.error(
-                "dir cache file format error '%s' [skipping the file]",
-                obj.path_info,
-            )
-            raise ObjectFormatError(f"{obj} is corrupted")
-
-        tree = cls.from_list(raw)
-        tree.path_info = obj.path_info
-        tree.fs = obj.fs
-        tree.hash_info = hash_info
-
-        return tree
-
-    def filter(self, odb, prefix):
-        obj = self._dict.get(prefix)
-        if obj:
-            return obj
-
-        depth = len(prefix)
-        tree = Tree(None, None, None)
-        try:
-            for key, obj in self.trie.items(prefix):
-                tree.add(key[depth:], obj)
-        except KeyError:
-            return None
-        tree.digest()
-        odb.add(tree.path_info, tree.fs, tree.hash_info)
-        return tree
-
-
 def save(odb, obj, **kwargs):
+    from .tree import Tree
+
     if isinstance(obj, Tree):
         for _, entry in Tqdm(obj):
             odb.add(entry.path_info, entry.fs, entry.hash_info, **kwargs)
@@ -206,6 +79,8 @@ def save(odb, obj, **kwargs):
 
 
 def check(odb, obj):
+    from .tree import Tree
+
     odb.check(obj.hash_info)
 
     if isinstance(obj, Tree):
@@ -214,82 +89,11 @@ def check(odb, obj):
 
 
 def load(odb, hash_info):
+    from .tree import Tree
+
     if hash_info.isdir:
         return Tree.load(odb, hash_info)
     return odb.get(hash_info)
-
-
-def _get_dir_size(odb, tree):
-    try:
-        return sum(
-            odb.fs.getsize(odb.hash_to_path_info(obj.hash_info.value))
-            for _, obj in tree
-        )
-    except FileNotFoundError:
-        return None
-
-
-def _diff(ancestor, other, allow_removed=False):
-    from dictdiffer import diff
-
-    from dvc.exceptions import MergeError
-
-    allowed = ["add"]
-    if allow_removed:
-        allowed.append("remove")
-
-    result = list(diff(ancestor, other))
-    for typ, _, _ in result:
-        if typ not in allowed:
-            raise MergeError(
-                "unable to auto-merge directories with diff that contains "
-                f"'{typ}'ed files"
-            )
-    return result
-
-
-def _merge(ancestor, our, their):
-    import copy
-
-    from dictdiffer import patch
-
-    our_diff = _diff(ancestor, our)
-    if not our_diff:
-        return copy.deepcopy(their)
-
-    their_diff = _diff(ancestor, their)
-    if not their_diff:
-        return copy.deepcopy(our)
-
-    # make sure there are no conflicting files
-    _diff(our, their, allow_removed=True)
-
-    return patch(our_diff + their_diff, ancestor)
-
-
-def merge(odb, ancestor_info, our_info, their_info):
-    assert our_info
-    assert their_info
-
-    if ancestor_info:
-        ancestor = load(odb, ancestor_info)
-    else:
-        ancestor = Tree(None, None, None)
-
-    our = load(odb, our_info)
-    their = load(odb, their_info)
-
-    merged_dict = _merge(ancestor.as_dict(), our.as_dict(), their.as_dict(),)
-
-    merged = Tree(None, None, None)
-    for key, hi in merged_dict.items():
-        merged.add(key, hi)
-    merged.digest()
-
-    odb.add(merged.path_info, merged.fs, merged.hash_info)
-    hash_info = merged.hash_info
-    hash_info.size = _get_dir_size(odb, merged)
-    return hash_info
 
 
 def _transfer_file(odb, from_fs, from_info):
@@ -342,6 +146,8 @@ def _transfer_directory_contents(odb, from_fs, from_info, jobs, pbar):
 
 
 def _transfer_directory(odb, from_fs, from_info, jobs, no_progress_bar=False):
+    from .tree import Tree
+
     tree = Tree(None, None, None)
 
     with Tqdm(total=1, unit="Files", disable=no_progress_bar) as pbar:
